+++ include/rapidjson/error/en.h
@@ -116,6 +116,9 @@ inline const RAPIDJSON_ERROR_CHARTYPE* GetValidateError_En(ValidateErrorCode val
         case kValidateErrorShape:                       return RAPIDJSON_ERROR_STRING("Property has a shape that does not match the schema.");
         case kValidateErrorPythonImport:                return RAPIDJSON_ERROR_STRING("Property is not an importable Python object.");
         case kValidateErrorInvalidSchema:               return RAPIDJSON_ERROR_STRING("Property is not a valid JSON schema.");
+        case kNormalizeErrorAliasDuplicate:             return RAPIDJSON_ERROR_STRING("Aliased property already exists in the normalized document.");
+        case kNormalizeErrorCircularAlias:              return RAPIDJSON_ERROR_STRING("Property has circular aliases.");
+        case kNormalizeErrorConflictingAliases:         return RAPIDJSON_ERROR_STRING("Property has conflicting aliases.");
 #endif // RAPIDJSON_YGGDRASIL
         default:                                        return RAPIDJSON_ERROR_STRING("Unknown error.");
     }